httpFetcher now uses promise internally and reorganized Utils
